Finish everything
+ Configure Babel to understand React (JSX) for unit test
+ Install tools supporting babel
+ Refactoring functions that makes them testable for the test cases
+ Update my todo list (mini Kanban)

P.s.: In README, I was told "I can request the data from https://" but
not have to. And sure I've tested my solution with the right now payload
from SpaceX that that link resolved. But I think it's still not a huge
deal if or not in my implementation do that.